+++ src/services/tests/mod.rs
@@ -1,4 +1,4 @@
 mod invalidate_caches;
 mod parse_id;
 mod transactions_history;
-mod transactions_queue;
+mod transactions_queued;
